11060: create tomselect helper module
